+++ src/swap/defi/abi/WRAPPED_FTM_ABI.js
@@ -1,3 +1,4 @@
+// ABI for wrapped tokens
 export default [
   {
     inputs: [],
